pyproject.toml
[project]
name = "fast-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.8.1",
    "python-dotenv>=1.1.0",
    "redis>=5.0.0",
    "psycopg>=3.1.0",
]
